requirements.txt
tensorflow-cpu==2.13.0
numpy<=1.24.3
pandas==2.0.3
matplotlib==3.7.2
gradio==3.50.2
fastapi<0.100
pydantic<2
typing-extensions==4.5.0
